+++ src/pages/SwapV2/index.tsx
@@ -40,7 +40,7 @@ import ProgressSteps from 'components/ProgressSteps'
 import { SwitchLocaleLink } from 'components/SwitchLocaleLink'
 import { INITIAL_ALLOWED_SLIPPAGE } from 'constants/index'
 import { useActiveWeb3React } from 'hooks'
-import { useAllTokens, useCurrency, searchInactiveToken } from 'hooks/Tokens'
+import { useAllTokens, useCurrency } from 'hooks/Tokens'
 import { ApprovalState, useApproveCallbackFromTradeV2 } from 'hooks/useApproveCallback'
 import useWrapCallback, { WrapType } from 'hooks/useWrapCallback'
 import { useToggleTransactionSettingsMenu, useWalletModalToggle } from 'state/application/hooks'
@@ -86,14 +86,14 @@ import TopTrendingSoonTokensInCurrentNetwork from 'components/TopTrendingSoonTok
 import { clientData } from 'constants/clientData'
 import { NETWORKS_INFO, SUPPORTED_NETWORKS } from 'constants/networks'
 import { useActiveNetwork } from 'hooks/useActiveNetwork'
-import { convertSymbol, convertToSlug, getNetworkSlug, getSymbolSlug, checkPairInWhiteList } from 'utils/string'
+import { convertToSlug, getNetworkSlug, getSymbolSlug } from 'utils/string'
+import { checkPairInWhiteList, convertSymbol } from 'utils/tokenInfo'
 import { filterTokensWithExactKeyword } from 'components/SearchModal/filtering'
 import { useRef } from 'react'
 import { nativeOnChain } from 'constants/tokens'
 
 import Footer from 'components/Footer/Footer'
 import usePrevious from 'hooks/usePrevious'
-import { useAllLists, useInactiveListUrls } from 'state/lists/hooks'
 enum ACTIVE_TAB {
   SWAP,
   INFO,
@@ -143,7 +143,6 @@ export default function Swap({ history }: RouteComponentProps) {
   const [activeTab, setActiveTab] = useState<ACTIVE_TAB>(ACTIVE_TAB.SWAP)
 
   const loadedUrlParams = useDefaultsFromURLSearch()
-  const [tokensUrlNotInDefault, setTokensUrlNotInDefault] = useState<Token[]>([])
 
   // token warning stuff
   const [loadedInputCurrency, loadedOutputCurrency] = [
@@ -237,7 +236,6 @@ export default function Swap({ history }: RouteComponentProps) {
   // reset if they close warning without tokens in params
   const handleDismissTokenWarning = useCallback(() => {
     setDismissTokenWarning(true)
-    setTokensUrlNotInDefault([])
   }, [])
 
   // modal and loading
@@ -402,8 +400,6 @@ export default function Swap({ history }: RouteComponentProps) {
   const navigate = (url: string) => {
     history.push(`${url}${window.location.search}`) // keep query params
   }
-  const lists = useAllLists()
-  const inactiveUrls = useInactiveListUrls()
 
   function findTokenPairFromUrl() {
     if (!refIsCheckNetworkAutoSelect.current || refIsImportUserToken.current || !Object.keys(defaultTokens).length)
@@ -452,19 +448,6 @@ export default function Swap({ history }: RouteComponentProps) {
     const toToken = findToken(toCurrency)
 
     if (!toToken || !fromToken) {
-      const { isInWhiteList } = checkPairInWhiteList(chainId, fromCurrency, toCurrency)
-      if (isInWhiteList && !dismissTokenWarning) {
-        // token in whitelist but not import yet => show popup import
-        const params = [chainId, defaultTokens, lists, inactiveUrls] as const
-        const unActiveTokens = [
-          fromToken ? undefined : searchInactiveToken(fromCurrency, ...params),
-          toToken ? undefined : searchInactiveToken(toCurrency, ...params),
-        ].filter(Boolean)
-        if (unActiveTokens.length) {
-          setTokensUrlNotInDefault(unActiveTokens as Token[])
-          return
-        }
-      }
       navigate('/swap')
       return
     }
@@ -584,13 +567,12 @@ export default function Swap({ history }: RouteComponentProps) {
     checkPairInWhiteList(chainId, getSymbolSlug(currencyIn), getSymbolSlug(currencyOut)).isInWhiteList
 
   const [actualShowTokenInfo, setActualShowTokenInfo] = useState(true)
-  const isShowImportModal =
-    (tokensUrlNotInDefault.length > 0 || importTokensNotInDefault.length > 0) && !dismissTokenWarning
+
   return (
     <>
       <TokenWarningModal
-        isOpen={isShowImportModal}
-        tokens={tokensUrlNotInDefault.length ? tokensUrlNotInDefault : importTokensNotInDefault}
+        isOpen={importTokensNotInDefault.length > 0 && !dismissTokenWarning}
+        tokens={importTokensNotInDefault}
         onConfirm={handleDismissTokenWarning}
         onDismiss={handleDismissTokenWarning}
       />
